services/deployment-pipeline/pyproject.toml
[project]
name = "deployment-pipeline"
version = "0.1.0"
description = "Deployment Pipeline — automated build, test, and deploy orchestration for ARCHITECT"
requires-python = ">=3.12"
dependencies = [
    "architect-common",
    "architect-events",
    "architect-observability",
    "temporalio>=1.9",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deployment_pipeline"]

[tool.uv.sources]
architect-common = { workspace = true }
architect-events = { workspace = true }
architect-observability = { workspace = true }
